Backend/requirements.txt
# Existing packages (kept as is)
Flask==2.3.3
Flask-CORS==4.0.0
Flask-SQLAlchemy==3.0.5
Flask-JWT-Extended==4.5.3
psycopg2-binary==2.9.9
python-dotenv==1.0.0
bcrypt==4.1.2
Werkzeug==2.3.7
gunicorn==21.2.0
SQLAlchemy==2.0.23
google-auth==2.29.0
google-auth-oauthlib==1.2.0
google-auth-httplib2==0.2.0
google-api-python-client==2.125.0

# NEW: WebSocket and real-time dependencies
Flask-SocketIO==5.3.4
python-socketio==5.9.0
eventlet==0.33.3

# NEW: Additional utilities (if not already installed)
reportlab==4.0.4
python-docx==0.8.11
Pillow==10.0.0